Repository level abstraction changes
