+++ plugins/crypto/licenses/commons-logging-1.2.jar.sha1
@@ -1,0 +1,1 @@
+4bfc12adfe4842bf07b657f0369c4cb522955686
