+++ plugins/crypto/licenses/commons-logging-NOTICE.txt
@@ -1,0 +1,5 @@
+Apache Commons Logging
+Copyright 2003-2016 The Apache Software Foundation
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
