+++ plugins/crypto/licenses/slf4j-api-1.7.36.jar.sha1
@@ -1,0 +1,1 @@
+6c62681a2f655b49963a5983b8b0950a6120ae14
