--- plugins/crypto/sample-extension/resources/META-INF/services/org.opensearch.crypto.spi.CryptoKeyProviderExtension
@@ -1,6 +1,0 @@
-#
-# Copyright OpenSearch Contributors
-# SPDX-License-Identifier: Apache-2.0
-#
-
-org.opensearch.crypto.sampleextension.SampleExtensionPlugin
