+++ plugins/crypto/sample-extension/src/main/resources/org/opensearch/cryptoplugin/sampleextension/raw_key
@@ -1,0 +1,1 @@
+[57, 59, -48, -8, -44, 9, -78, 16, 106, -80, 66, -41, 66, 43, -88, 7, 47, -23, -16, -43, 99, 104, -8, -74, 46, -117, -111, -41, -39, -69, 5, 117]
