--- plugins/crypto/src/main/java/org/opensearch/crypto/cache/package-info.java
@@ -1,13 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * Caching layer for generated data keys
- */
-package org.opensearch.crypto.cache;
