--- plugins/crypto/src/main/java/org/opensearch/crypto/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * Crypto plugin to for encryption and decryption use cases.
- */
-package org.opensearch.crypto;
