+++ plugins/crypto/src/main/java/org/opensearch/cryptoplugin/keyprovider/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Key provider package
+ */
+package org.opensearch.cryptoplugin.keyprovider;
