+++ plugins/crypto/src/test/resources/org/opensearch/cryptoplugin/encrypted_key
