+++ plugins/crypto/src/test/resources/org/opensearch/cryptoplugin/raw_key
